Refactor for DRY code in callbacks
